Allow the use of generated images as inputs

# Conflicts:
#	google/generativeai/vision_models/_vision_models.py
